[1.78.x][canary] Brockit uplift
This PR uplifts all the content of brockit up to today, as there has
been quite a long development cycle for it, that has not been captured
in ohter branch outside `master`.
